requirements.txt
pydantic>=2.0.0
torch>=1.13.0
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
transformers>=4.21.0
sentence-transformers>=2.2.0
pymorphy3>=1.0.0
razdel>=0.5.0
PyYAML>=6.0
colorlog>=6.7.0
nltk>=3.8.0
